add ignore save empty note

+++ app/src/main/java/com/notes/crud/zenolab/crud/ScrollFragment.java
@@ -4,13 +4,19 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.AsyncTask;
+
 import androidx.annotation.Nullable;
+
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
+
 import androidx.fragment.app.Fragment;
+
 import android.os.Bundle;
+
 import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -20,7 +26,7 @@ import android.widget.Toast;
 
 import com.notes.crud.zenolab.crud.adapter.RecyclerTouchListener;
 import com.notes.crud.zenolab.crud.adapter.RecyclerViewClickListener;
-import com.notes.crud.zenolab.crud.adapter.ViewAdapter;
+import com.notes.crud.zenolab.crud.adapter.NoteAdapter;
 import com.notes.crud.zenolab.crud.room.entity.Note;
 import com.notes.crud.zenolab.crud.room.room_database.NoteDatabase;
 
@@ -31,7 +37,7 @@ import java.util.List;
 /**
  * A placeholder fragment containing a simple view.
  */
-public class ScrollFragment extends Fragment  {
+public class ScrollFragment extends Fragment {
 
     private static final String LOG_TAG = new RuntimeException().getStackTrace()[0].getClassName();
 
@@ -41,16 +47,16 @@ public class ScrollFragment extends Fragment  {
 
     private TextView textViewMsg;
     private NoteDatabase noteDatabase;
-    private ViewAdapter notesAdapter;
+    private NoteAdapter notesAdapter;
     FloatingActionButton fab;
     private int pos;
+    AttentionDialogFragment attentionDialogFragment;
 
 
+    private View rootView;
 
-
-    private  View rootView;
-
-    public ScrollFragment() { }
+    public ScrollFragment() {
+    }
 
 
     @Override
@@ -64,43 +70,42 @@ public class ScrollFragment extends Fragment  {
         return initViews(inflater.inflate(R.layout.scroll_main, container, false));
     }
 
-    private void displayList(){
-        //DATABASE
+    private void displayList() {
         noteDatabase = NoteDatabase.getInstance(getActivity());
         new RetrieveTask(this).execute();
     }
 
 
-    private View initViews(View view){
+    private View initViews(View view) {
         context = view.getContext();
-        textViewMsg =  (TextView) view.findViewById(R.id.tv__empty);
+        textViewMsg = (TextView) view.findViewById(R.id.tv__empty);
 
-        fab = (FloatingActionButton) view.findViewById(R.id.fab);
+        fab = ((FloatingActionButton) getActivity().findViewById(R.id.fab_center));
         fab.setOnClickListener(listener);
 
         recycler = view.findViewById(R.id.recyclerView);
         recycler.setLayoutManager(new LinearLayoutManager(context));
         notes = new ArrayList<>();
-        notesAdapter = new ViewAdapter(context,notes);
-        initListClickListener( recycler );
+        notesAdapter = new NoteAdapter(context, notes);
+        initListClickListener(recycler);
         recycler.setAdapter(notesAdapter);
         displayList();
 
         return view;
     }
 
-    //----------------------------------------------------------------------------------------------
-    private void initListClickListener(RecyclerView recyclerView ) {
+    private void initListClickListener(RecyclerView recyclerView) {
         recyclerView.addOnItemTouchListener(new RecyclerTouchListener(getActivity(),
                 recyclerView, new RecyclerViewClickListener() {
             @Override
             public void onClick(View view, final int position) {
                 startActivity(new Intent(getActivity(), AddNoteActivity.class).
-                                putExtra("note",notes.get(position)));
+                        putExtra("note", notes.get(position)));
             }
+
             @Override
             public void onLongClick(View view, int position) {
-                action( position);
+                action(position);
             }
         }));
     }
@@ -112,52 +117,63 @@ public class ScrollFragment extends Fragment  {
                 .setItems(new String[]{"Delete", "Update"}, new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialogInterface, int i) {
-                        switch (i){
+                        switch (i) {
                             case 0:
-                                noteDatabase.getNoteDao().deleteNote(notes.get(position));
-                                notes.remove(position);
-                                listVisibility();
+                                showDialogFragment(position);
                                 break;
                             case 1:
                                 try {
                                     getActivity().startActivityForResult(
                                             new Intent(getActivity(),
-                                                    AddNoteActivity.class).putExtra("note",notes.get(position)),
+                                                    AddNoteActivity.class)
+                                                    .putExtra("note", notes.get(position)),
                                             100);
                                     break;
-                                }catch (NullPointerException npe){
-                                    Toast.makeText(getActivity(), "Screen loading error", Toast.LENGTH_SHORT).show();
+                                } catch (NullPointerException npe) {
+                                    Toast.makeText(getActivity(),
+                                            "Screen loading error", Toast.LENGTH_SHORT).show();
                                 }
                         }
                     }
                 }).show();
-
     }
-    //----------------------------------------------------------------------------------------------
 
-    private View.OnClickListener listener = new View.OnClickListener() {
-        @Override
-        public void onClick(View view) {
-          //  startActivityForResult(new Intent(NoteListActivity.this,AddNoteActivity.class),100);
-            startActivity(new Intent(getActivity(), AddNoteActivity.class));
-        }
+
+    public void showDialogFragment(int position) {
+
+        attentionDialogFragment = new AttentionDialogFragment();
+        attentionDialogFragment.registerCallBack(new PredicateListener() {
+            @Override
+            public void predicateListener(boolean arg) {
+                if (arg) {
+                    noteDatabase.getNoteDao().deleteNote(notes.get(position));
+                    notes.remove(position);
+                    listVisibility();
+                }
+            }
+        });
+        attentionDialogFragment.show(getFragmentManager(), "custom");
+    }
+
+    private View.OnClickListener listener = view -> {
+        startActivity(new Intent(getActivity(), AddNoteActivity.class));
     };
 
     @Override
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == 100 && resultCode > 0 ){
-            if( resultCode == 1){
+        if (requestCode == 100 && resultCode > 0) {
+            if (resultCode == 1) {
                 notes.add((Note) data.getSerializableExtra("note"));
-            }else if( resultCode == 2){
-                notes.set(pos,(Note) data.getSerializableExtra("note"));
+            } else if (resultCode == 2) {
+                notes.set(pos, (Note) data.getSerializableExtra("note"));
             }
             listVisibility();
         }
     }
 
-    private void listVisibility(){
+    private void listVisibility() {
         int emptyMsgVisibility = View.GONE;
-        if (notes.size() == 0){ // no item to display
+        if (notes.size() == 0) { // no item to display
             if (textViewMsg.getVisibility() == View.GONE)
                 emptyMsgVisibility = View.VISIBLE;
         }
@@ -179,36 +195,35 @@ public class ScrollFragment extends Fragment  {
 
     }
 
-    //==============================================================================================
-    private static class RetrieveTask extends AsyncTask<Void,Void,List<Note>> {
 
-        private WeakReference<ScrollFragment> activityReference;
+    private static class RetrieveTask extends AsyncTask<Void, Void, List<Note>> {
+
+        private WeakReference<ScrollFragment> scrollFragmentWeakReference;
 
         // only retain a weak reference to the activity
         RetrieveTask(ScrollFragment scrollFragment) {
-            activityReference = new WeakReference<>(scrollFragment);
+            scrollFragmentWeakReference = new WeakReference<>(scrollFragment);
         }
 
         @Override
         protected List<Note> doInBackground(Void... voids) {
-            if (activityReference.get()!=null)
-                return activityReference.get().noteDatabase.getNoteDao().getNotes();
+            if (scrollFragmentWeakReference.get() != null)
+                return scrollFragmentWeakReference.get().noteDatabase.getNoteDao().getNotes();
             else
                 return null;
         }
 
         @Override
         protected void onPostExecute(List<Note> notes) {
-            if (notes!=null && notes.size()>0 ){
-                activityReference.get().notes.clear();
-                activityReference.get().notes.addAll(notes);
+            if (notes != null && notes.size() > 0) {
+                scrollFragmentWeakReference.get().notes.clear();
+                scrollFragmentWeakReference.get().notes.addAll(notes);
                 // hides empty text view
-                activityReference.get().textViewMsg.setVisibility(View.GONE);
-                activityReference.get().notesAdapter.notifyDataSetChanged();
-
-                Log.d(LOG_TAG,"...");
+                scrollFragmentWeakReference.get().textViewMsg.setVisibility(View.GONE);
+                scrollFragmentWeakReference.get().notesAdapter.notifyDataSetChanged();
+                Log.d(LOG_TAG, "...");
             }
         }
     }
-    ///=============================================================================================
+
 }
